pyproject.toml
[project]
name = "discord-event-registrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "google-api-python-client>=2.167.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "python-dotenv>=1.1.0",
    "returns>=0.25.0",
]

[dependency-groups]
dev = [
    "google-api-python-client-stubs>=1.29.0",
]
